Extract getPaper to own module
I suspect other addons will want this in the future :)
